modife int->double

+++ Projet-brick-breaker-main copie/brick.cc
@@ -7,7 +7,7 @@
 using namespace std;
 
 
-Brick :: Brick (int t, double x, double y, double c, int h ):
+Brick :: Brick (double t, double x, double y, double c, double h ):
 t(t),x(x),y(y),c(c),h(h)
 {}
